requirements.txt
Django==1.11.8
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.1
pytz==2018.4
